refactor: extract insight breakdown pie chart
